Common functionality moved to common project
