+++ officefloor/bom/pom.xml
@@ -427,6 +427,11 @@
 			</dependency>
 			<dependency>
 				<groupId>net.officefloor.cabinet</groupId>
+				<artifactId>officecabinet_common</artifactId>
+				<version>${revision}</version>
+			</dependency>
+			<dependency>
+				<groupId>net.officefloor.cabinet</groupId>
 				<artifactId>officecabinet_cosmosdb</artifactId>
 				<version>${revision}</version>
 			</dependency>
