+++ officefloor/cabinet/officecabinet_common/.gitignore
@@ -1,0 +1,1 @@
+/target/
