+++ officefloor/cabinet/officecabinet_common/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>net.officefloor.cabinet</groupId>
+		<artifactId>cabinet</artifactId>
+		<version>${revision}</version>
+	</parent>
+	<artifactId>officecabinet_common</artifactId>
+	<name>OfficeFloor Filing Cabinet Common</name>
+	<description>Cloud storage common functionality for OfficeFloor</description>
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	</properties>
+	<dependencies>
+		<dependency>
+			<groupId>net.officefloor.cabinet</groupId>
+			<artifactId>officecabinet</artifactId>
+		</dependency>
+	</dependencies>
+</project>
