+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/CabinetUtil.java
