+++ officefloor/cabinet/officecabinet_firestore/pom.xml
@@ -16,7 +16,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>net.officefloor.cabinet</groupId>
-			<artifactId>officecabinet</artifactId>
+			<artifactId>officecabinet_common</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>net.officefloor.persistence</groupId>
