+++ officefloor/cabinet/officecabinet_firestore/src/test/java/net/officefloor/cabinet/firestore/FirestoreOfficeCabinetTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import net.officefloor.cabinet.AbstractOfficeCabinetTest;
 import net.officefloor.cabinet.AttributeTypesDocument;
-import net.officefloor.cabinet.OfficeCabinetArchive;
+import net.officefloor.cabinet.spi.OfficeCabinetArchive;
 import net.officefloor.nosql.firestore.test.FirestoreExtension;
 import net.officefloor.test.UsesDockerTest;
