+++ officefloor/cabinet/officecabinet_test/src/main/java/net/officefloor/cabinet/AbstractOfficeCabinetTest.java
@@ -26,6 +26,8 @@ import static org.junit.jupiter.api.Assertions.assertSame;
 
 import org.junit.jupiter.api.Test;
 
+import net.officefloor.cabinet.spi.OfficeCabinetArchive;
+
 /**
  * Tests Office Cabinet.
  *
