+++ officefloor/cabinet/pom.xml
@@ -15,6 +15,7 @@
 	<modules>
 		<module>officecabinet</module>
 		<module>officecabinet_test</module>
+		<module>officecabinet_common</module>
 		<module>officecabinet_dynamodb</module>
 		<module>officecabinet_firestore</module>
 		<module>officecabinet_cosmosdb</module>
